mctsragg/requirements.txt
faiss-cpu
sentence-transformers
requests
pandas
tqdm
numpy